Configure renovate to only handle security updates
Renovate generates a lot of noise with its default configuration
because it opens a dependency update PR for every single major,
minor, or patch version of every dependency.

This updates the renovate configuration to only handle security
updates so that we can reduce PR noise and be more intentional
about dependency updates.  Configuration here was taken from:
https://github.com/renovatebot/renovate/discussions/15490